requirements.txt
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
certifi==2020.12.5
Django==3.1.6
fonttools==4.25.0
inflection==0.5.1
jsonpointer==2.1
jsonschema
munkres==1.1.4
openpyxl==3.0.10
patsy==0.5.3
pillow==10.3.0
ply==3.11
protobuf==3.20.3
py4j==0.10.6
modules==0.2.8
PyDispatcher==2.0.5
spyder==0.4.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
pyspark==2.3.0
jsonrpc==1.0.0
pytz==2021.1
pyviz_comms
sqlparse==0.4.1
webencodings==0.5.1